requirements.txt
# pip installs
gradio
lingua-language-detector
faiss-cpu
opencv-python
folium
requests-mock
##xformers<0.0.26
accelerate==0.33
geopy
jupyter
pandas
plotly
tokenizers
torchvision==0.17.1
#torch==2.3.1+cu121
tqdm
transformers
numpy==1.26.4
# git installs
unsloth[kaggle-new]@ git+https://github.com/unslothai/unsloth.git